[Refactor] Site cleanup and SEO improvements (#8)
* Code cleanup, SEO improvements

* Moved meta tags to different component
